feat(inference): cycle detector
