+++ crates/biome_module_graph/tests/spec_tests.rs
@@ -7,14 +7,9 @@ use std::fs::read_link;
 use std::ops::Deref;
 use std::sync::Arc;
 
-use biome_db::ParsedSource;
-use biome_db::testing::{
-    Events, assert_function_query_was_not_run, assert_function_query_was_run,
-    function_query_will_execute_position,
-};
 use biome_resolver::ResolveError;
 
-use crate::snap::{ModuleGraphSnapshot, source_files_from_memory_fs, write_source_file};
+use crate::snap::ModuleGraphSnapshot;
 use biome_configuration::{Configuration, HtmlConfiguration};
 use biome_css_parser::{CssModulesKind, CssParserOptions};
 use biome_deserialize::json::deserialize_from_json_str;
@@ -22,29 +17,18 @@ use biome_fs::{BiomePath, FileSystem, MemoryFileSystem, OsFileSystem, normalize_
 use biome_html_parser::HtmlParserOptions;
 use biome_js_semantic::ScopeId;
 use biome_js_syntax::AnyJsRoot;
-use biome_js_type_info::{
-    TypeData, TypeResolver, format_inferred_type,
-    interned_types::{
-        FunctionParameter as InferredFunctionParameter, InternedInterface as InferredInterface,
-        InternedMergedReference as InferredMergedReference, InternedUnion as InferredUnion,
-        Literal as InferredLiteral, ReturnType as InferredReturnType, TypeData as InferredTypeData,
-        TypeMemberKind as InferredTypeMemberKind,
-    },
-};
+use biome_js_type_info::{TypeData, TypeResolver};
 use biome_json_parser::{JsonParserOptions, parse_json};
 use biome_json_value::{JsonObject, JsonString};
 use biome_languages::css::{CssEmbeddingKind, EmbeddingHtmlKind, EmbeddingStyleApplicability};
 use biome_languages::{CssFileSource, DocumentFileSource, HtmlFileSource, JsFileSource};
 use biome_module_graph::{
-    CallExpressionTypeInput, HtmlEmbeddedContent, ImportSymbol, InferredModuleTypes, JsExport,
-    JsImport, JsImportPath, JsImportPhase, JsModuleInfoDiagnostic, JsOwnExport, JsReexport,
-    ModuleDb, ModuleDiagnostic, ModuleInfo, ModuleInfoKind, ModuleResolver, NormalizeTypeInput,
-    PathInfoCache, ResolvedPath, SymbolFromModuleInfo, find_js_exported_symbol,
-    infer_call_expression_type as infer_call_expression_type_query, infer_module_types,
-    infer_module_types_bottom_up, is_class_referenced_by_importers,
-    normalize_type as normalize_type_query, resolve_css_module, resolve_html_module,
-    resolve_js_module, transitive_importers_of, traverse_import_tree_for_classes,
-    traverse_import_tree_for_html_classes,
+    HtmlEmbeddedContent, ImportSymbol, JsExport, JsImport, JsImportPath, JsImportPhase,
+    JsModuleInfoDiagnostic, JsOwnExport, JsReexport, ModuleDb, ModuleDiagnostic, ModuleInfo,
+    ModuleInfoKind, ModuleResolver, PathInfoCache, ResolvedPath, SymbolFromModuleInfo,
+    find_js_exported_symbol, is_class_referenced_by_importers, resolve_css_module,
+    resolve_html_module, resolve_js_module, transitive_importers_of,
+    traverse_import_tree_for_classes, traverse_import_tree_for_html_classes,
 };
 use biome_package::{Dependencies, PackageJson};
 use biome_project_layout::ProjectLayout;
@@ -57,350 +41,7 @@ use biome_test_utils::{get_added_js_paths, get_css_added_paths};
 use biome_workspace_db::WorkspaceDb;
 use camino::{Utf8Path, Utf8PathBuf};
 use rustc_hash::FxHashSet;
-use salsa::Storage;
 use walkdir::WalkDir;
-
-#[salsa::db]
-struct TestModuleDb {
-    modules: BTreeMap<Utf8PathBuf, ModuleInfo>,
-    events: Events,
-    storage: Storage<Self>,
-}
-
-impl TestModuleDb {
-    fn new() -> Self {
-        let events = Events::default();
-        Self {
-            modules: BTreeMap::new(),
-            storage: salsa::Storage::new(Some(Box::new({
-                let events = events.clone();
-                move |event| {
-                    events.0.lock().unwrap().push(event);
-                }
-            }))),
-            events,
-        }
-    }
-
-    fn take_salsa_events(&mut self) -> Vec<salsa::Event> {
-        std::mem::take(&mut *self.events.0.lock().unwrap())
-    }
-
-    fn clear_salsa_events(&mut self) {
-        self.take_salsa_events();
-    }
-}
-
-#[salsa::db]
-impl salsa::Database for TestModuleDb {}
-
-#[salsa::db]
-impl biome_db::Db for TestModuleDb {
-    fn parsed_source_for_path(&self, _path: &Utf8Path) -> Option<ParsedSource> {
-        None
-    }
-}
-
-#[salsa::db]
-impl biome_module_graph::TypeDb for TestModuleDb {}
-
-#[salsa::db]
-impl ModuleDb for TestModuleDb {
-    fn module_for_path(&self, path: &Utf8Path) -> Option<ModuleInfo> {
-        self.modules.get(path).copied()
-    }
-
-    fn for_each_module(&self, f: &mut dyn FnMut(&Utf8Path, &ModuleInfoKind)) {
-        for (path, module) in &self.modules {
-            f(path, &module.kind(self));
-        }
-    }
-}
-
-#[salsa::tracked]
-fn inferred_expression_count(db: &dyn ModuleDb, module: ModuleInfo) -> usize {
-    infer_module_types(db, module).map_or(0, |inferred| inferred.expressions.len())
-}
-
-fn is_inferred_instance_of<'db>(
-    db: &'db dyn ModuleDb,
-    ty: InferredTypeData<'db>,
-    inner: InferredTypeData<'db>,
-) -> bool {
-    matches!(ty, InferredTypeData::InstanceOf(instance) if instance.ty(db) == inner)
-}
-
-fn is_inferred_string<'db>(db: &'db dyn ModuleDb, ty: InferredTypeData<'db>) -> bool {
-    ty == InferredTypeData::String
-        || is_inferred_instance_of(db, ty, InferredTypeData::String)
-        || matches!(ty, InferredTypeData::Literal(literal) if matches!(literal.literal(db), InferredLiteral::String(_)))
-}
-
-fn is_inferred_number<'db>(db: &'db dyn ModuleDb, ty: InferredTypeData<'db>) -> bool {
-    ty == InferredTypeData::Number
-        || is_inferred_instance_of(db, ty, InferredTypeData::Number)
-        || matches!(ty, InferredTypeData::Literal(literal) if matches!(literal.literal(db), InferredLiteral::Number(_)))
-}
-
-fn contains_inferred_string<'db>(db: &'db dyn ModuleDb, ty: InferredTypeData<'db>) -> bool {
-    if is_inferred_string(db, ty) {
-        return true;
-    }
-
-    match ty {
-        InferredTypeData::Union(union) => union
-            .types(db)
-            .iter()
-            .any(|ty| contains_inferred_string(db, *ty)),
-        _ => false,
-    }
-}
-
-fn contains_inferred_number<'db>(db: &'db dyn ModuleDb, ty: InferredTypeData<'db>) -> bool {
-    if is_inferred_number(db, ty) {
-        return true;
-    }
-
-    match ty {
-        InferredTypeData::Union(union) => union
-            .types(db)
-            .iter()
-            .any(|ty| contains_inferred_number(db, *ty)),
-        _ => false,
-    }
-}
-
-fn contains_inferred_undefined<'db>(db: &'db dyn ModuleDb, ty: InferredTypeData<'db>) -> bool {
-    if ty == InferredTypeData::Undefined {
-        return true;
-    }
-
-    match ty {
-        InferredTypeData::Union(union) => union
-            .types(db)
-            .iter()
-            .any(|ty| contains_inferred_undefined(db, *ty)),
-        _ => false,
-    }
-}
-
-fn is_inferred_promise_with_type_parameter<'db>(
-    db: &'db dyn ModuleDb,
-    ty: InferredTypeData<'db>,
-    predicate: impl Fn(InferredTypeData<'db>) -> bool,
-) -> bool {
-    let InferredTypeData::InstanceOf(instance) = ty else {
-        return false;
-    };
-
-    ty.is_promise_instance(db) && instance.type_parameters(db).iter().any(|ty| predicate(*ty))
-}
-
-fn assert_inferred_function_returns_string<'db>(db: &'db dyn ModuleDb, ty: InferredTypeData<'db>) {
-    let InferredTypeData::Function(function) = ty else {
-        panic!("type must be inferred as a function");
-    };
-    let InferredReturnType::Type(return_ty) = function.return_type(db) else {
-        panic!("function return type must be inferred as a type");
-    };
-
-    assert!(is_inferred_string(db, *return_ty));
-}
-
-fn assert_inferred_function_returns_number<'db>(db: &'db dyn ModuleDb, ty: InferredTypeData<'db>) {
-    let InferredTypeData::Function(function) = ty else {
-        panic!("type must be inferred as a function");
-    };
-    let InferredReturnType::Type(return_ty) = function.return_type(db) else {
-        panic!("function return type must be inferred as a type");
-    };
-
-    assert!(is_inferred_number(db, *return_ty));
-}
-
-fn local_type_id_of_instance<'db>(
-    db: &'db dyn ModuleDb,
-    ty: InferredTypeData<'db>,
-) -> Option<usize> {
-    let InferredTypeData::InstanceOf(instance) = ty else {
-        return None;
-    };
-    let InferredTypeData::Local(local) = instance.ty(db) else {
-        return None;
-    };
-
-    Some(local.type_id(db).index())
-}
-
-fn infer_call_expression_type<'db>(
-    db: &'db dyn ModuleDb,
-    module: ModuleInfo,
-    callee: InferredTypeData<'db>,
-    args: Vec<InferredTypeData<'db>>,
-) -> InferredTypeData<'db> {
-    infer_call_expression_type_query(
-        db,
-        CallExpressionTypeInput::new(db, module, callee, args.into_boxed_slice()),
-    )
-}
-
-fn normalize_type<'db>(
-    db: &'db dyn ModuleDb,
-    module: ModuleInfo,
-    ty: InferredTypeData<'db>,
-) -> InferredTypeData<'db> {
-    normalize_type_query(db, NormalizeTypeInput::new(db, module, ty))
-}
-
-fn interface_member_ty<'db>(
-    db: &'db dyn ModuleDb,
-    interface: InferredInterface<'db>,
-    member_name: &str,
-) -> Option<InferredTypeData<'db>> {
-    interface.members(db).iter().find_map(|member| {
-        matches!(
-            &member.kind,
-            InferredTypeMemberKind::Named(name) if name.text() == member_name,
-        )
-        .then_some(member.ty)
-    })
-}
-
-fn object_member_ty_by_name<'db>(
-    db: &'db dyn ModuleDb,
-    ty: InferredTypeData<'db>,
-    member_name: &str,
-) -> Option<(InferredTypeMemberKind<'db>, InferredTypeData<'db>)> {
-    let object = match ty {
-        InferredTypeData::Object(object) => object,
-        InferredTypeData::InstanceOf(instance) => match instance.ty(db) {
-            InferredTypeData::Object(object) => object,
-            _ => return None,
-        },
-        _ => return None,
-    };
-
-    object.members(db).iter().find_map(|member| {
-        member
-            .kind
-            .name()
-            .is_some_and(|name| name.text() == member_name)
-            .then(|| (member.kind.clone(), member.ty))
-    })
-}
-
-fn inferred_binding_ty_by_name<'db>(
-    db: &'db dyn ModuleDb,
-    module: ModuleInfo,
-    inferred: &InferredModuleTypes<'db>,
-    name: &str,
-) -> Option<InferredTypeData<'db>> {
-    let ModuleInfoKind::Js(info) = module.kind(db) else {
-        return None;
-    };
-    let binding = info.semantic_model.all_bindings().find(|binding| {
-        binding
-            .tree()
-            .name_token()
-            .is_ok_and(|token| token.text_trimmed() == name)
-    })?;
-
-    inferred
-        .binding_type_data
-        .get(&binding.syntax().text_trimmed_range())
-        .map(|data| data.ty)
-}
-
-fn inferred_function_return_ty_by_name<'db>(
-    db: &'db dyn ModuleDb,
-    module: ModuleInfo,
-    inferred: &InferredModuleTypes<'db>,
-    name: &str,
-) -> Option<InferredTypeData<'db>> {
-    let binding_ty = inferred_binding_ty_by_name(db, module, inferred, name)?;
-    let function = inferred
-        .resolve_type(db, binding_ty)
-        .callable_function(db)?;
-    let InferredReturnType::Type(return_ty) = function.return_type(db) else {
-        return None;
-    };
-
-    Some(inferred.resolve_type(db, *return_ty))
-}
-
-fn inferred_overload_ty_by_name<'db>(
-    db: &'db dyn ModuleDb,
-    module: ModuleInfo,
-    inferred: &InferredModuleTypes<'db>,
-    name: &str,
-) -> Option<InferredTypeData<'db>> {
-    let ModuleInfoKind::Js(info) = module.kind(db) else {
-        return None;
-    };
-
-    info.semantic_model
-        .all_bindings()
-        .filter(|binding| {
-            binding
-                .tree()
-                .name_token()
-                .is_ok_and(|token| token.text_trimmed() == name)
-        })
-        .filter_map(|binding| {
-            inferred
-                .binding_type_data
-                .get(&binding.syntax().text_trimmed_range())
-                .map(|data| inferred.resolve_type(db, data.ty))
-        })
-        .find(|ty| {
-            matches!(
-                ty,
-                InferredTypeData::Object(object)
-                    if object
-                        .members(db)
-                        .iter()
-                        .filter(|member| member.kind.is_call_signature())
-                        .count()
-                        >= 2
-            )
-        })
-}
-
-fn assert_inferred_type_snapshot(test_name: &str, db: &dyn ModuleDb, fs: &MemoryFileSystem) {
-    let mut content = String::new();
-    let files = source_files_from_memory_fs(fs);
-    for (file_name, source_code) in &files {
-        let file_name = Utf8PathBuf::from(file_name.as_str());
-        write_source_file(&mut content, &file_name, source_code, None);
-
-        let Some(module) = db.module_for_path(file_name.as_path()) else {
-            continue;
-        };
-        let Some(inferred) = infer_module_types(db, module) else {
-            continue;
-        };
-        if inferred.types.is_empty() {
-            continue;
-        }
-
-        content.push_str("\n\n## Inferred types\n\n```");
-        for (index, ty) in inferred.types.iter().enumerate() {
-            content.push_str("\nModule TypeId(");
-            content.push_str(&index.to_string());
-            content.push_str(") => ");
-            content.push_str(&format_inferred_type(db, *ty));
-            content.push('\n');
-        }
-        content.push_str("```\n");
-    }
-
-    insta::with_settings!({
-        snapshot_path => "snapshots",
-        prepend_module_to_snapshot => false,
-    }, {
-        insta::assert_snapshot!(test_name, content);
-    });
-}
 
 fn build_js_db(
     fs: &dyn biome_resolver::FsWithResolverProxy,
@@ -426,44 +67,6 @@ fn build_js_db(
             ModuleInfoKind::Js(module_info),
         );
         db.modules.pin().insert(path.as_path().to_path_buf(), md);
-    }
-    db
-}
-
-fn resolve_js_module_kind_for_test(
-    fs: &MemoryFileSystem,
-    path: &str,
-    infer_types: bool,
-) -> ModuleInfoKind {
-    let paths = [BiomePath::new(path)];
-    let mut added_paths = get_added_js_paths(fs, &paths);
-    let (path, root, semantic_model) = added_paths.pop().expect("module must parse");
-    let (module_info, _, _) = resolve_js_module(
-        root,
-        path,
-        fs,
-        &ProjectLayout::default(),
-        semantic_model,
-        &PathInfoCache::default(),
-        infer_types,
-    );
-
-    ModuleInfoKind::Js(module_info)
-}
-
-fn build_js_test_module_db(
-    fs: &MemoryFileSystem,
-    paths: &[&str],
-    infer_types: bool,
-) -> TestModuleDb {
-    let mut db = TestModuleDb::new();
-    for path in paths {
-        let module_info = ModuleInfo::new(
-            &db,
-            Utf8PathBuf::from(*path),
-            resolve_js_module_kind_for_test(fs, path, infer_types),
-        );
-        db.modules.insert(Utf8PathBuf::from(*path), module_info);
     }
     db
 }
@@ -988,2979 +591,6 @@ fn test_export_const_type_declaration_with_namespace() {
 
     let snapshot = ModuleGraphSnapshot::new(&db, &fs);
     snapshot.assert_snapshot("test_export_const_type_declaration_with_namespace");
-}
-
-#[test]
-fn test_infer_module_types_resolves_imported_exported_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/types.ts".into(),
-        r#"
-            export type Foo = string;
-        "#,
-    );
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            import type { Foo } from "./types.ts";
-
-            export const value: Foo = "value";
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/types.ts", "/src/index.ts"], true);
-    let types_module = db
-        .module_for_path(Utf8Path::new("/src/types.ts"))
-        .expect("module must exist");
-    let inferred_types = infer_module_types(&db, types_module).expect("types must be inferred");
-    let ModuleInfoKind::Js(types_info) = types_module.kind(&db) else {
-        panic!("module must be JavaScript");
-    };
-    assert_eq!(inferred_types.types.len(), types_info.raw_types.len());
-
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    assert!(
-        inferred
-            .binding_type_data
-            .values()
-            .any(|data| data.ty == InferredTypeData::String)
-    );
-}
-
-#[test]
-fn test_infer_module_types_resolves_nested_function_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/types.ts".into(),
-        r#"
-            export type Foo = string;
-        "#,
-    );
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            import type { Foo } from "./types.ts";
-
-            export function value(): Foo {
-                return "value";
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/types.ts", "/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    assert!(inferred.types.iter().any(|ty| {
-        matches!(
-            *ty,
-            InferredTypeData::Function(function)
-                if matches!(
-                    function.return_type(&db),
-                    InferredReturnType::Type(return_ty)
-                        if is_inferred_instance_of(&db, *return_ty, InferredTypeData::String),
-                )
-        )
-    }));
-}
-
-#[test]
-fn test_infer_module_types_resolves_nested_object_member_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/types.ts".into(),
-        r#"
-            export type Foo = string;
-        "#,
-    );
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            import type { Foo } from "./types.ts";
-
-            export type Boxed = {
-                value: Foo;
-            };
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/types.ts", "/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    assert!(inferred.types.iter().any(|ty| {
-        matches!(
-            *ty,
-            InferredTypeData::Object(object)
-                if object.members(&db).iter().any(|member| {
-                    matches!(
-                        &member.kind,
-                        InferredTypeMemberKind::Named(name) if name.text() == "value",
-                    ) && is_inferred_instance_of(&db, member.ty, InferredTypeData::String)
-                })
-        )
-    }));
-}
-
-#[test]
-fn test_infer_module_types_resolves_nested_union_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/types.ts".into(),
-        r#"
-            export type Foo = string;
-            export type Bar = number;
-        "#,
-    );
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            import type { Foo, Bar } from "./types.ts";
-
-            export type Value = Foo | Bar;
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/types.ts", "/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    assert!(inferred.types.iter().any(|ty| {
-        matches!(
-            *ty,
-            InferredTypeData::Union(union)
-                if union
-                    .types(&db)
-                    .iter()
-                    .any(|ty| is_inferred_instance_of(&db, *ty, InferredTypeData::String))
-                    && union
-                        .types(&db)
-                        .iter()
-                        .any(|ty| is_inferred_instance_of(&db, *ty, InferredTypeData::Number))
-        )
-    }));
-}
-
-#[test]
-fn test_infer_module_types_uses_local_handle_for_recursive_class_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export class Foo {
-                name: string;
-
-                static create(): Foo {
-                    return new Foo();
-                }
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let (class_index, class) = inferred
-        .types
-        .iter()
-        .enumerate()
-        .find_map(|(index, ty)| match ty {
-            InferredTypeData::Class(class)
-                if class
-                    .name(&db)
-                    .as_ref()
-                    .is_some_and(|name| name.text() == "Foo") =>
-            {
-                Some((index, class))
-            }
-            _ => None,
-        })
-        .expect("Foo class type must be inferred");
-
-    let return_ty = class
-        .members(&db)
-        .iter()
-        .find_map(|member| {
-            if !matches!(
-                &member.kind,
-                InferredTypeMemberKind::NamedStatic(name) if name.text() == "create",
-            ) {
-                return None;
-            }
-
-            let InferredTypeData::Function(function) = member.ty else {
-                return None;
-            };
-            match function.return_type(&db) {
-                InferredReturnType::Type(return_ty) => Some(*return_ty),
-                InferredReturnType::Predicate(_) | InferredReturnType::Asserts(_) => None,
-            }
-        })
-        .expect("Foo.create return type must be inferred");
-
-    let InferredTypeData::InstanceOf(instance) = return_ty else {
-        panic!("Foo.create must return an instance type");
-    };
-    let InferredTypeData::Local(local) = instance.ty(&db) else {
-        panic!("Foo.create must return a local handle to Foo");
-    };
-
-    assert_eq!(local.module(&db), inferred.module_key);
-    assert_eq!(local.type_id(&db).index(), class_index);
-
-    let name_ty = inferred
-        .find_member_type(&db, return_ty, "name")
-        .expect("Foo.create().name must be inferred");
-    assert!(is_inferred_string(&db, name_ty));
-
-    let create_ty = inferred
-        .find_member_type(&db, InferredTypeData::Local(local), "create")
-        .expect("Foo.create must be found through the local handle");
-    assert!(matches!(create_ty, InferredTypeData::Function(_)));
-}
-
-#[test]
-fn test_infer_module_types_uses_local_handles_for_recursive_interfaces() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export interface A {
-                b: B;
-            }
-
-            export interface B {
-                a: A;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let (a_index, a_interface) = inferred
-        .types
-        .iter()
-        .enumerate()
-        .find_map(|(index, ty)| match ty {
-            InferredTypeData::Interface(interface) if interface.name(&db).text() == "A" => {
-                Some((index, *interface))
-            }
-            _ => None,
-        })
-        .expect("A interface type must be inferred");
-    let (b_index, b_interface) = inferred
-        .types
-        .iter()
-        .enumerate()
-        .find_map(|(index, ty)| match ty {
-            InferredTypeData::Interface(interface) if interface.name(&db).text() == "B" => {
-                Some((index, *interface))
-            }
-            _ => None,
-        })
-        .expect("B interface type must be inferred");
-
-    let a_b_ty = interface_member_ty(&db, a_interface, "b").expect("A.b must be inferred");
-    let b_a_ty = interface_member_ty(&db, b_interface, "a").expect("B.a must be inferred");
-
-    assert_eq!(local_type_id_of_instance(&db, a_b_ty), Some(b_index));
-    assert_eq!(local_type_id_of_instance(&db, b_a_ty), Some(a_index));
-
-    let b_from_a = inferred
-        .find_member_type(&db, b_a_ty, "b")
-        .expect("B.a.b must be inferred through A's local handle");
-    let a_from_b = inferred
-        .find_member_type(&db, a_b_ty, "a")
-        .expect("A.b.a must be inferred through B's local handle");
-
-    assert_eq!(local_type_id_of_instance(&db, b_from_a), Some(b_index));
-    assert_eq!(local_type_id_of_instance(&db, a_from_b), Some(a_index));
-}
-
-#[test]
-fn test_infer_module_types_resolves_inherited_class_members() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export class Base {
-                name: string;
-
-                static label(): string {
-                    return "base";
-                }
-            }
-
-            export class Derived extends Base {
-                value: number;
-            }
-
-            export const derived: Derived = new Derived();
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let derived_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "derived")
-        .expect("derived binding type must be inferred");
-    let name_ty = inferred
-        .find_member_type(&db, derived_ty, "name")
-        .expect("Derived instance must inherit Base.name");
-    assert!(is_inferred_string(&db, name_ty));
-
-    let derived_class_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "Derived")
-        .expect("Derived class type must be inferred");
-    let label_ty = inferred
-        .find_member_type(&db, derived_class_ty, "label")
-        .expect("Derived class must inherit Base.label");
-    assert!(matches!(label_ty, InferredTypeData::Function(_)));
-}
-
-#[test]
-fn test_infer_module_types_resolves_inherited_interface_members() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export interface Base {
-                name: string;
-            }
-
-            export interface Derived extends Base {
-                value: number;
-            }
-
-            export const derived: Derived = {
-                name: "derived",
-                value: 1,
-            };
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let derived_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "derived")
-        .expect("derived binding type must be inferred");
-    let name_ty = inferred
-        .find_member_type(&db, derived_ty, "name")
-        .expect("Derived interface must inherit Base.name");
-    assert!(is_inferred_string(&db, name_ty));
-}
-
-#[test]
-fn test_infer_module_types_resolves_intersection_members() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            type WithName = {
-                name: string;
-            };
-
-            type WithValue = {
-                value: number;
-            };
-
-            export const combined: WithName & WithValue = {
-                name: "combined",
-                value: 1,
-            };
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let combined_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "combined")
-        .expect("combined binding type must be inferred");
-
-    let name_ty = inferred
-        .find_member_type(&db, combined_ty, "name")
-        .expect("intersection must expose WithName.name");
-    assert!(is_inferred_string(&db, name_ty));
-
-    let value_ty = inferred
-        .find_member_type(&db, combined_ty, "value")
-        .expect("intersection must expose WithValue.value");
-    assert!(is_inferred_number(&db, value_ty));
-}
-
-#[test]
-fn test_infer_module_types_resolves_union_member_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            type StringValue = {
-                value: string;
-            };
-
-            type NumberValue = {
-                value: number;
-            };
-
-            export const item: StringValue | NumberValue = {
-                value: "item",
-            };
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let item_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "item")
-        .expect("item binding type must be inferred");
-    let value_ty = inferred
-        .find_member_type(&db, item_ty, "value")
-        .expect("union must expose shared value member");
-
-    assert!(matches!(
-        value_ty,
-        InferredTypeData::Union(union)
-            if union
-                .types(&db)
-                .iter()
-                .any(|ty| is_inferred_string(&db, *ty))
-                && union
-                    .types(&db)
-                    .iter()
-                    .any(|ty| is_inferred_number(&db, *ty))
-    ));
-}
-
-#[test]
-fn test_infer_module_types_resolves_generic_constraint_members() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function readName<T extends { name: string }>(value: T): T {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let read_name_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "readName")
-        .expect("readName binding type must be inferred");
-    let InferredTypeData::Function(read_name) = inferred.resolve_type(&db, read_name_ty) else {
-        panic!("readName must be inferred as a function");
-    };
-    let value_ty = read_name
-        .parameters(&db)
-        .iter()
-        .find_map(|parameter| match parameter {
-            InferredFunctionParameter::Named(parameter) if parameter.name.text() == "value" => {
-                Some(parameter.ty)
-            }
-            _ => None,
-        })
-        .expect("value parameter type must be inferred");
-
-    let name_ty = inferred
-        .find_member_type(&db, value_ty, "name")
-        .expect("generic constraint must expose name");
-    assert!(is_inferred_string(&db, name_ty));
-}
-
-#[test]
-fn test_infer_module_types_resolves_string_index_signature_members() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            type Dictionary = {
-                [key: string]: number;
-            };
-
-            export const dictionary: Dictionary = {};
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let dictionary_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "dictionary")
-        .expect("dictionary binding type must be inferred");
-
-    let value_ty = inferred
-        .find_member_type(&db, dictionary_ty, "anything")
-        .expect("string index signature must expose arbitrary string members");
-    assert!(is_inferred_number(&db, value_ty));
-}
-
-#[test]
-fn test_infer_module_types_resolves_computed_string_literal_members() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export const object = {
-                ["name"]: "object",
-            };
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let object_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "object")
-        .expect("object binding type must be inferred");
-
-    let name_ty = inferred
-        .find_member_type(&db, object_ty, "name")
-        .expect("computed string literal member must match its literal name");
-    assert!(is_inferred_string(&db, name_ty));
-
-    assert!(inferred.find_member_type(&db, object_ty, "other").is_none());
-}
-
-#[test]
-fn test_infer_module_types_resolves_merged_reference_members() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            interface Foo {
-                typeName: string;
-            }
-
-            const Foo = {
-                valueName: 1,
-            };
-
-            export { Foo };
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let ModuleInfoKind::Js(js_info) = index_module.kind(&db) else {
-        panic!("module must be JavaScript");
-    };
-    let foo_ty = match js_info.exports.get("Foo").and_then(JsExport::as_own_export) {
-        Some(JsOwnExport::Type(resolved_id)) => inferred
-            .types
-            .get(resolved_id.index())
-            .copied()
-            .expect("Foo export type must be inferred"),
-        _ => panic!("Foo export must have a type"),
-    };
-    let type_name_ty = inferred
-        .find_member_type(&db, foo_ty, "typeName")
-        .expect("merged type side must expose Foo.typeName");
-    assert!(is_inferred_string(&db, type_name_ty));
-
-    let value_name_ty = inferred
-        .find_member_type(&db, foo_ty, "valueName")
-        .expect("merged value side must expose Foo.valueName");
-    assert!(is_inferred_number(&db, value_name_ty));
-}
-
-#[test]
-fn test_normalize_type_collapses_equal_merged_reference_targets() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            type Foo = {
-                name: "foo";
-            };
-
-            const Foo = {
-                name: "foo",
-            } as const;
-
-            export { Foo };
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let ModuleInfoKind::Js(js_info) = index_module.kind(&db) else {
-        panic!("module must be JavaScript");
-    };
-    let foo_ty = match js_info.exports.get("Foo").and_then(JsExport::as_own_export) {
-        Some(JsOwnExport::Type(resolved_id)) => inferred
-            .types
-            .get(resolved_id.index())
-            .copied()
-            .expect("Foo export type must be inferred"),
-        _ => panic!("Foo export must have a type"),
-    };
-
-    assert!(
-        matches!(foo_ty, InferredTypeData::MergedReference(_)),
-        "Foo export must be a merged reference, got {foo_ty:?}"
-    );
-
-    let InferredTypeData::MergedReference(reference) = foo_ty else {
-        panic!("Foo export must be a merged reference, got {foo_ty:?}");
-    };
-    let target = reference
-        .targets(&db)
-        .next()
-        .expect("merged reference must have a target");
-    let expected_ty = normalize_type(&db, index_module, target);
-    let duplicated_target_ty = InferredTypeData::MergedReference(InferredMergedReference::new(
-        &db,
-        Some(target),
-        Some(target),
-        None,
-    ));
-    let normalized_ty = normalize_type(&db, index_module, duplicated_target_ty);
-    assert_eq!(normalized_ty, expected_ty);
-
-    let name_ty = inferred
-        .find_member_type(&db, normalized_ty, "name")
-        .expect("normalized merged reference must expose Foo.name");
-    assert!(is_inferred_string(&db, name_ty));
-
-    assert_inferred_type_snapshot(
-        "test_normalize_type_collapses_equal_merged_reference_targets",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_normalize_type_resolves_typeof_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            const source = {
-                name: "source",
-            } as const;
-
-            type Source = typeof source;
-
-            export function readSource(value: Source): Source {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let source_ty = inferred_function_return_ty_by_name(&db, index_module, &inferred, "readSource")
-        .expect("readSource return type must be inferred");
-    let normalized_ty = normalize_type(&db, index_module, source_ty);
-    let InferredTypeData::Object(_) = normalized_ty else {
-        panic!("typeof type must normalize to its target object, got {normalized_ty:?}");
-    };
-
-    let name_ty = inferred
-        .find_member_type(&db, normalized_ty, "name")
-        .expect("normalized typeof type must expose source.name");
-    assert!(is_inferred_string(&db, name_ty));
-
-    assert_inferred_type_snapshot("test_normalize_type_resolves_typeof_type", &db, &fs);
-}
-
-#[test]
-fn test_normalize_type_preserves_recursive_local_edge() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            type Tree = number | Promise<Tree>;
-
-            export function readTree(value: Tree): Tree {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let tree_ty = inferred_function_return_ty_by_name(&db, index_module, &inferred, "readTree")
-        .expect("readTree return type must be inferred");
-    let tree_index = local_type_id_of_instance(&db, tree_ty)
-        .expect("readTree must return an instance of the local Tree type");
-    let normalized_ty = normalize_type(&db, index_module, tree_ty);
-
-    let InferredTypeData::Union(union) = normalized_ty else {
-        panic!("recursive Tree type must normalize to a union, got {normalized_ty:?}");
-    };
-    let normalized_tree = format_inferred_type(&db, normalized_ty);
-    assert!(
-        union
-            .types(&db)
-            .iter()
-            .any(|ty| is_inferred_number(&db, *ty)),
-        "recursive Tree union must keep its number branch"
-    );
-    assert!(
-        union.types(&db).iter().any(|ty| {
-            matches!(
-                ty,
-                InferredTypeData::InstanceOf(instance)
-                    if instance.type_parameters(&db).iter().any(|parameter| {
-                        matches!(
-                            parameter,
-                            InferredTypeData::Local(local)
-                                if local.type_id(&db).index() == tree_index
-                        )
-                        || local_type_id_of_instance(&db, *parameter) == Some(tree_index)
-                    })
-            )
-        }),
-        "recursive Tree union must keep the recursive local edge: {normalized_tree}"
-    );
-
-    assert_inferred_type_snapshot(
-        "test_normalize_type_preserves_recursive_local_edge",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_normalize_type_preserves_recursive_array_local_edge() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            type Tree = number | Tree[];
-
-            export function readTree(value: Tree): Tree {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let tree_ty = inferred_function_return_ty_by_name(&db, index_module, &inferred, "readTree")
-        .expect("readTree return type must be inferred");
-    let tree_index = local_type_id_of_instance(&db, tree_ty)
-        .expect("readTree must return an instance of the local Tree type");
-    let normalized_ty = normalize_type(&db, index_module, tree_ty);
-
-    let InferredTypeData::Union(union) = normalized_ty else {
-        panic!("recursive Tree type must normalize to a union, got {normalized_ty:?}");
-    };
-    let normalized_tree = format_inferred_type(&db, normalized_ty);
-    assert!(
-        union
-            .types(&db)
-            .iter()
-            .any(|ty| is_inferred_number(&db, *ty)),
-        "recursive Tree union must keep its number branch"
-    );
-    assert!(
-        union.types(&db).iter().any(|ty| {
-            matches!(
-                ty,
-                InferredTypeData::InstanceOf(instance)
-                    if matches!(
-                        instance.ty(&db),
-                        InferredTypeData::Class(class)
-                            if class
-                                .name(&db)
-                                .as_ref()
-                                .is_some_and(|name| name.text() == "Array")
-                    )
-                    && instance.type_parameters(&db).iter().any(|parameter| {
-                        matches!(
-                            parameter,
-                            InferredTypeData::Local(local)
-                                if local.type_id(&db).index() == tree_index
-                        )
-                        || local_type_id_of_instance(&db, *parameter) == Some(tree_index)
-                    })
-            )
-        }),
-        "recursive Tree union must keep the recursive Array local edge: {normalized_tree}"
-    );
-
-    assert_inferred_type_snapshot(
-        "test_normalize_type_preserves_recursive_array_local_edge",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_module_types_resolves_generic_builtin_instances_on_build() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function readMap(value: Map<string, number>): Map<string, number> {
-                return value;
-            }
-
-            export function readSet(value: Set<string>): Set<string> {
-                return value;
-            }
-
-            export function readWeakMap(value: WeakMap<object, string>): WeakMap<object, string> {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let map_ty = inferred_function_return_ty_by_name(&db, index_module, &inferred, "readMap")
-        .expect("readMap return type must be inferred");
-    let InferredTypeData::InstanceOf(map_instance) = map_ty else {
-        panic!("readMap must return a Map instance, got {map_ty:?}");
-    };
-    let InferredTypeData::Class(map_class) = map_instance.ty(&db) else {
-        panic!("readMap must return a class instance");
-    };
-    assert_eq!(map_class.name(&db).as_ref().map(Text::text), Some("Map"));
-    assert_eq!(map_instance.type_parameters(&db).len(), 2);
-    assert!(is_inferred_string(
-        &db,
-        map_instance.type_parameters(&db)[0]
-    ));
-    assert!(is_inferred_number(
-        &db,
-        map_instance.type_parameters(&db)[1]
-    ));
-
-    let set_ty = inferred_function_return_ty_by_name(&db, index_module, &inferred, "readSet")
-        .expect("readSet return type must be inferred");
-    let InferredTypeData::InstanceOf(set_instance) = set_ty else {
-        panic!("readSet must return a Set instance, got {set_ty:?}");
-    };
-    let InferredTypeData::Class(set_class) = set_instance.ty(&db) else {
-        panic!("readSet must return a class instance");
-    };
-    assert_eq!(set_class.name(&db).as_ref().map(Text::text), Some("Set"));
-    assert_eq!(set_instance.type_parameters(&db).len(), 1);
-    assert!(is_inferred_string(
-        &db,
-        set_instance.type_parameters(&db)[0]
-    ));
-
-    let weak_map_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readWeakMap")
-            .expect("readWeakMap return type must be inferred");
-    let InferredTypeData::InstanceOf(weak_map_instance) = weak_map_ty else {
-        panic!("readWeakMap must return a WeakMap instance, got {weak_map_ty:?}");
-    };
-    let InferredTypeData::Class(weak_map_class) = weak_map_instance.ty(&db) else {
-        panic!("readWeakMap must return a class instance");
-    };
-    assert_eq!(
-        weak_map_class.name(&db).as_ref().map(Text::text),
-        Some("WeakMap")
-    );
-    assert_eq!(weak_map_instance.type_parameters(&db).len(), 2);
-    assert!(is_inferred_string(
-        &db,
-        weak_map_instance.type_parameters(&db)[1]
-    ));
-
-    assert_inferred_type_snapshot(
-        "test_infer_module_types_resolves_generic_builtin_instances_on_build",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_module_types_resolves_builtin_global_identities_on_build() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function readRegExp(value: RegExp): RegExp {
-                return value;
-            }
-
-            export function readDate(value: Date): Date {
-                return value;
-            }
-
-            export function readError(value: Error): Error {
-                return value;
-            }
-
-            export function readSymbol(value: Symbol): Symbol {
-                return value;
-            }
-
-            export function readDisposable(value: Disposable): Disposable {
-                return value;
-            }
-
-            export function readAsyncDisposable(value: AsyncDisposable): AsyncDisposable {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    for (function_name, class_name) in [
-        ("readRegExp", "RegExp"),
-        ("readDate", "Date"),
-        ("readError", "Error"),
-        ("readSymbol", "Symbol"),
-    ] {
-        let ty = inferred_function_return_ty_by_name(&db, index_module, &inferred, function_name)
-            .unwrap_or_else(|| panic!("{function_name} return type must be inferred"));
-        let InferredTypeData::InstanceOf(instance) = ty else {
-            panic!("{function_name} must return a {class_name} instance, got {ty:?}");
-        };
-        let InferredTypeData::Class(class) = instance.ty(&db) else {
-            panic!("{function_name} must return a {class_name} instance, got {ty:?}");
-        };
-        assert_eq!(class.name(&db).as_ref().map(Text::text), Some(class_name));
-    }
-
-    for (function_name, interface_name) in [
-        ("readDisposable", "Disposable"),
-        ("readAsyncDisposable", "AsyncDisposable"),
-    ] {
-        let ty = inferred_function_return_ty_by_name(&db, index_module, &inferred, function_name)
-            .unwrap_or_else(|| panic!("{function_name} return type must be inferred"));
-        let InferredTypeData::InstanceOf(instance) = ty else {
-            panic!("{function_name} must return an {interface_name} instance, got {ty:?}");
-        };
-        let InferredTypeData::Interface(interface) = instance.ty(&db) else {
-            panic!("{function_name} must return an {interface_name} instance, got {ty:?}");
-        };
-        assert_eq!(interface.name(&db).text(), interface_name);
-    }
-
-    assert_inferred_type_snapshot(
-        "test_infer_module_types_resolves_builtin_global_identities_on_build",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_module_types_resolves_record_index_signature_on_build() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function readRecord(value: Record<string, number>): Record<string, number> {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let record_ty = inferred_function_return_ty_by_name(&db, index_module, &inferred, "readRecord")
-        .expect("readRecord return type must be inferred");
-    let InferredTypeData::InstanceOf(record_instance) = record_ty else {
-        panic!("readRecord must return an object instance, got {record_ty:?}");
-    };
-    let InferredTypeData::Object(object) = record_instance.ty(&db) else {
-        panic!("readRecord must return an object instance, got {record_ty:?}");
-    };
-    assert_eq!(object.members(&db).len(), 1);
-
-    let item_ty = inferred
-        .find_member_type(&db, record_ty, "item")
-        .expect("Record<string, number> must expose a number string index signature");
-    assert!(is_inferred_number(&db, item_ty));
-
-    assert_inferred_type_snapshot(
-        "test_infer_module_types_resolves_record_index_signature_on_build",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_module_types_resolves_utility_type_members_on_build() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            type Source = {
-                [key: string]: string | number | undefined;
-                name: string;
-                value: number;
-                optional?: string;
-            };
-
-            export function readPick(
-                value: Pick<Source, "name" | "value">,
-            ): Pick<Source, "name" | "value"> {
-                return value;
-            }
-
-            export function readOmit(
-                value: Omit<Source, "value">,
-            ): Omit<Source, "value"> {
-                return value;
-            }
-
-            export function readPartial(value: Partial<Source>): Partial<Source> {
-                return value;
-            }
-
-            export function readRequired(value: Required<Source>): Required<Source> {
-                return value;
-            }
-
-            export function readReadonly(value: Readonly<Source>): Readonly<Source> {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let pick_ty = inferred_function_return_ty_by_name(&db, index_module, &inferred, "readPick")
-        .expect("readPick return type must be inferred");
-    let (_, pick_name_ty) =
-        object_member_ty_by_name(&db, pick_ty, "name").expect("Pick<Source, ...> must keep name");
-    let (_, pick_value_ty) =
-        object_member_ty_by_name(&db, pick_ty, "value").expect("Pick<Source, ...> must keep value");
-    assert!(is_inferred_string(&db, pick_name_ty));
-    assert!(is_inferred_number(&db, pick_value_ty));
-    assert!(object_member_ty_by_name(&db, pick_ty, "optional").is_none());
-    assert!(
-        inferred
-            .find_member_type(&db, pick_ty, "anything")
-            .is_none()
-    );
-
-    let omit_ty = inferred_function_return_ty_by_name(&db, index_module, &inferred, "readOmit")
-        .expect("readOmit return type must be inferred");
-    let (_, omit_name_ty) =
-        object_member_ty_by_name(&db, omit_ty, "name").expect("Omit<Source, ...> must keep name");
-    assert!(is_inferred_string(&db, omit_name_ty));
-    assert!(object_member_ty_by_name(&db, omit_ty, "value").is_none());
-    let omit_index_ty = inferred
-        .find_member_type(&db, omit_ty, "anything")
-        .expect("Omit<Source, ...> must preserve the string index signature");
-    assert!(contains_inferred_string(&db, omit_index_ty));
-    assert!(contains_inferred_number(&db, omit_index_ty));
-
-    let partial_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readPartial")
-            .expect("readPartial return type must be inferred");
-    let (partial_name_kind, partial_name_ty) =
-        object_member_ty_by_name(&db, partial_ty, "name").expect("Partial<Source> must keep name");
-    assert!(partial_name_kind.is_optional());
-    assert!(contains_inferred_string(&db, partial_name_ty));
-    assert!(contains_inferred_undefined(&db, partial_name_ty));
-
-    let required_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readRequired")
-            .expect("readRequired return type must be inferred");
-    let (required_optional_kind, required_optional_ty) =
-        object_member_ty_by_name(&db, required_ty, "optional")
-            .expect("Required<Source> must keep optional");
-    assert!(!required_optional_kind.is_optional());
-    assert!(is_inferred_string(&db, required_optional_ty));
-    assert!(!contains_inferred_undefined(&db, required_optional_ty));
-
-    let readonly_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readReadonly")
-            .expect("readReadonly return type must be inferred");
-    let (_, readonly_name_ty) = object_member_ty_by_name(&db, readonly_ty, "name")
-        .expect("Readonly<Source> must keep name");
-    assert!(is_inferred_string(&db, readonly_name_ty));
-
-    assert_inferred_type_snapshot(
-        "test_infer_module_types_resolves_utility_type_members_on_build",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_module_types_resolves_anonymous_default_class_export() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export default class {
-                name: string;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let ModuleInfoKind::Js(js_info) = index_module.kind(&db) else {
-        panic!("module must be JavaScript");
-    };
-    let default_ty = match js_info
-        .exports
-        .get("default")
-        .and_then(JsExport::as_own_export)
-    {
-        Some(JsOwnExport::Type(resolved_id)) => inferred
-            .types
-            .get(resolved_id.index())
-            .copied()
-            .expect("default export type must be inferred"),
-        _ => panic!("default export must have a type"),
-    };
-
-    assert!(!js_info.raw_types.is_empty());
-    assert!(inferred.named_type_ids.is_empty());
-
-    let name_ty = inferred
-        .find_member_type(&db, default_ty, "name")
-        .expect("anonymous default class member must be inferred");
-    assert!(is_inferred_string(&db, name_ty));
-}
-
-#[test]
-fn test_infer_module_types_resolves_anonymous_default_function_export() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export default function(): string {
-                return "value";
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let ModuleInfoKind::Js(js_info) = index_module.kind(&db) else {
-        panic!("module must be JavaScript");
-    };
-    let default_ty = match js_info
-        .exports
-        .get("default")
-        .and_then(JsExport::as_own_export)
-    {
-        Some(JsOwnExport::Type(resolved_id)) => inferred
-            .types
-            .get(resolved_id.index())
-            .copied()
-            .expect("default export type must be inferred"),
-        _ => panic!("default export must have a type"),
-    };
-
-    assert!(!js_info.raw_types.is_empty());
-    assert!(inferred.named_type_ids.is_empty());
-    assert_inferred_function_returns_string(&db, inferred.resolve_type(&db, default_ty));
-}
-
-#[test]
-fn test_infer_module_types_resolves_imported_anonymous_default_class_members() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/base.ts".into(),
-        r#"
-            export default class {
-                name: string;
-            }
-        "#,
-    );
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            import Base from "./base.ts";
-
-            export const value: Base = new Base();
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/base.ts", "/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let value_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "value")
-        .expect("value binding type must be inferred");
-
-    let name_ty = inferred
-        .find_member_type(&db, value_ty, "name")
-        .expect("imported anonymous default class member must be inferred");
-    assert!(is_inferred_string(&db, name_ty));
-}
-
-#[test]
-fn test_infer_module_types_resolves_imported_anonymous_default_function() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/base.ts".into(),
-        r#"
-            export default function(): string {
-                return "value";
-            }
-        "#,
-    );
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            import readValue from "./base.ts";
-
-            export { readValue };
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/base.ts", "/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let read_value_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "readValue")
-        .expect("readValue binding type must be inferred");
-
-    assert_inferred_function_returns_string(&db, inferred.resolve_type(&db, read_value_ty));
-}
-
-#[test]
-fn test_infer_call_expression_type_resolves_imported_default_function_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/base.ts".into(),
-        r#"
-            export default function(): string {
-                return "value";
-            }
-        "#,
-    );
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            import readValue from "./base.ts";
-
-            export const value = readValue();
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/base.ts", "/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let read_value_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "readValue")
-        .expect("readValue binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, read_value_ty),
-        Vec::new(),
-    );
-
-    assert!(is_inferred_string(&db, call_ty));
-}
-
-#[test]
-fn test_infer_call_expression_type_resolves_annotated_function_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export const readValue: () => string = () => "value";
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let read_value_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "readValue")
-        .expect("readValue binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, read_value_ty),
-        Vec::new(),
-    );
-
-    assert!(is_inferred_string(&db, call_ty));
-}
-
-#[test]
-fn test_infer_call_expression_type_resolves_callable_interface_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export interface Reader {
-                (): string;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let reader_ty = inferred
-        .types
-        .iter()
-        .copied()
-        .find(|ty| {
-            matches!(
-                ty,
-                InferredTypeData::Interface(interface)
-                    if interface.name(&db).text() == "Reader"
-                        && interface
-                            .members(&db)
-                            .iter()
-                            .any(|member| member.kind.is_call_signature())
-            )
-        })
-        .expect("Reader interface type must be inferred");
-    let call_ty = infer_call_expression_type(&db, index_module, reader_ty, Vec::new());
-
-    assert!(is_inferred_string(&db, call_ty));
-}
-
-#[test]
-fn test_infer_call_expression_type_selects_callable_interface_overload_by_arity() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export interface Reader {
-                (): string;
-                (value: number): number;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let reader_ty = inferred
-        .types
-        .iter()
-        .copied()
-        .find(|ty| {
-            matches!(
-                ty,
-                InferredTypeData::Interface(interface)
-                    if interface.name(&db).text() == "Reader"
-                        && interface
-                            .members(&db)
-                            .iter()
-                            .filter(|member| member.kind.is_call_signature())
-                            .count()
-                            == 2
-            )
-        })
-        .expect("Reader interface type must be inferred");
-
-    let zero_arg_ty = infer_call_expression_type(&db, index_module, reader_ty, Vec::new());
-    assert!(is_inferred_string(&db, zero_arg_ty));
-
-    let one_arg_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        reader_ty,
-        Vec::from([InferredTypeData::Number]),
-    );
-    assert!(is_inferred_number(&db, one_arg_ty));
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_selects_callable_interface_overload_by_arity",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_call_expression_type_selects_callable_interface_overload_with_optional_parameter_by_arity()
- {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export interface Reader {
-                (): string;
-                (value?: number): number;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let reader_ty = inferred
-        .types
-        .iter()
-        .copied()
-        .find(|ty| {
-            matches!(
-                ty,
-                InferredTypeData::Interface(interface)
-                    if interface.name(&db).text() == "Reader"
-                        && interface
-                            .members(&db)
-                            .iter()
-                            .filter(|member| member.kind.is_call_signature())
-                            .count()
-                            == 2
-            )
-        })
-        .expect("Reader interface type must be inferred");
-
-    let zero_arg_ty = infer_call_expression_type(&db, index_module, reader_ty, Vec::new());
-    assert!(is_inferred_string(&db, zero_arg_ty));
-
-    let one_arg_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        reader_ty,
-        Vec::from([InferredTypeData::Number]),
-    );
-    assert!(is_inferred_number(&db, one_arg_ty));
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_selects_callable_interface_overload_with_optional_parameter_by_arity",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_call_expression_type_selects_callable_interface_overload_with_rest_parameter_by_arity()
- {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export interface Reader {
-                (): string;
-                (...values: number[]): number;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let reader_ty = inferred
-        .types
-        .iter()
-        .copied()
-        .find(|ty| {
-            matches!(
-                ty,
-                InferredTypeData::Interface(interface)
-                    if interface.name(&db).text() == "Reader"
-                        && interface
-                            .members(&db)
-                            .iter()
-                            .filter(|member| member.kind.is_call_signature())
-                            .count()
-                            == 2
-            )
-        })
-        .expect("Reader interface type must be inferred");
-
-    let zero_arg_ty = infer_call_expression_type(&db, index_module, reader_ty, Vec::new());
-    assert!(is_inferred_string(&db, zero_arg_ty));
-
-    let many_arg_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        reader_ty,
-        Vec::from([InferredTypeData::Number, InferredTypeData::Number]),
-    );
-    assert!(is_inferred_number(&db, many_arg_ty));
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_selects_callable_interface_overload_with_rest_parameter_by_arity",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_call_expression_type_resolves_callable_object_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export type Reader = {
-                (): string;
-            };
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let reader_ty = inferred
-        .types
-        .iter()
-        .copied()
-        .find(|ty| {
-            matches!(
-                ty,
-                InferredTypeData::Object(object)
-                    if object
-                        .members(&db)
-                        .iter()
-                        .any(|member| member.kind.is_call_signature())
-            )
-        })
-        .expect("Reader object type must be inferred");
-    let call_ty = infer_call_expression_type(&db, index_module, reader_ty, Vec::new());
-
-    assert!(is_inferred_string(&db, call_ty));
-}
-
-#[test]
-fn test_infer_call_expression_type_selects_callable_object_overload_by_arity() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export type Reader = {
-                (): string;
-                (value: number): number;
-            };
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let reader_ty = inferred
-        .types
-        .iter()
-        .copied()
-        .find(|ty| {
-            matches!(
-                ty,
-                InferredTypeData::Object(object)
-                    if object
-                        .members(&db)
-                        .iter()
-                        .filter(|member| member.kind.is_call_signature())
-                        .count()
-                        == 2
-            )
-        })
-        .expect("Reader object type must be inferred");
-
-    let zero_arg_ty = infer_call_expression_type(&db, index_module, reader_ty, Vec::new());
-    assert!(is_inferred_string(&db, zero_arg_ty));
-
-    let one_arg_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        reader_ty,
-        Vec::from([InferredTypeData::Number]),
-    );
-    assert!(is_inferred_number(&db, one_arg_ty));
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_selects_callable_object_overload_by_arity",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_call_expression_type_selects_function_declaration_overload_by_callback_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function bestEffort<T>(cb: () => Promise<T>): Promise<T | undefined>;
-            export function bestEffort<T>(cb: () => T): T | undefined;
-            export function bestEffort<T>(cb: (() => T) | (() => Promise<T>)) {
-                return cb();
-            }
-
-            export function readPromise(): Promise<string> {
-                return Promise.resolve("value");
-            }
-
-            export function readString(): string {
-                return "value";
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let best_effort_ty = inferred_overload_ty_by_name(&db, index_module, &inferred, "bestEffort")
-        .expect("bestEffort overload type must be inferred");
-    let read_promise_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "readPromise")
-        .expect("readPromise binding type must be inferred");
-    let read_string_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "readString")
-        .expect("readString binding type must be inferred");
-
-    let promise_result_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        best_effort_ty,
-        Vec::from([inferred.resolve_type(&db, read_promise_ty)]),
-    );
-    assert!(
-        promise_result_ty.is_promise_instance(&db),
-        "promise callback overload must return a Promise, got {promise_result_ty:?}",
-    );
-    let InferredTypeData::InstanceOf(instance) = promise_result_ty else {
-        panic!("promise callback overload must return a Promise instance");
-    };
-    assert!(
-        instance
-            .type_parameters(&db)
-            .iter()
-            .any(|ty| contains_inferred_string(&db, *ty)),
-        "promise callback overload must substitute the callback return type"
-    );
-
-    let sync_result_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        best_effort_ty,
-        Vec::from([inferred.resolve_type(&db, read_string_ty)]),
-    );
-    assert_ne!(sync_result_ty, InferredTypeData::Unknown);
-    assert!(
-        !sync_result_ty.is_promise_instance(&db),
-        "sync callback overload must not return a Promise, got {sync_result_ty:?}",
-    );
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_selects_function_declaration_overload_by_callback_return_type",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_call_expression_type_selects_imported_function_declaration_overload_by_callback_return_type()
- {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/functions.ts".into(),
-        r#"
-            export function bestEffort<T>(cb: () => Promise<T>): Promise<T | undefined>;
-            export function bestEffort<T>(cb: () => T): T | undefined;
-            export function bestEffort<T>(cb: (() => T) | (() => Promise<T>)) {
-                return cb();
-            }
-        "#,
-    );
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            import { bestEffort } from "./functions.ts";
-
-            export function readPromise(): Promise<string> {
-                return Promise.resolve("value");
-            }
-
-            export function readString(): string {
-                return "value";
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/functions.ts", "/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let best_effort_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "bestEffort")
-        .expect("bestEffort import type must be inferred");
-    let read_promise_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "readPromise")
-        .expect("readPromise binding type must be inferred");
-    let read_string_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "readString")
-        .expect("readString binding type must be inferred");
-    let best_effort_ty = inferred.resolve_type(&db, best_effort_ty);
-
-    let promise_result_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        best_effort_ty,
-        Vec::from([inferred.resolve_type(&db, read_promise_ty)]),
-    );
-    assert!(
-        promise_result_ty.is_promise_instance(&db),
-        "promise callback overload must return a Promise, got {promise_result_ty:?}",
-    );
-    let InferredTypeData::InstanceOf(instance) = promise_result_ty else {
-        panic!("promise callback overload must return a Promise instance");
-    };
-    assert!(
-        instance
-            .type_parameters(&db)
-            .iter()
-            .any(|ty| contains_inferred_string(&db, *ty)),
-        "promise callback overload must substitute the callback return type"
-    );
-
-    let sync_result_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        best_effort_ty,
-        Vec::from([inferred.resolve_type(&db, read_string_ty)]),
-    );
-    assert_ne!(sync_result_ty, InferredTypeData::Unknown);
-    assert!(
-        !sync_result_ty.is_promise_instance(&db),
-        "sync callback overload must not return a Promise, got {sync_result_ty:?}",
-    );
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_selects_imported_function_declaration_overload_by_callback_return_type",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_call_expression_type_substitutes_direct_generic_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function identity<T>(value: T): T {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let identity_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "identity")
-        .expect("identity binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, identity_ty),
-        Vec::from([InferredTypeData::Number]),
-    );
-
-    assert!(is_inferred_number(&db, call_ty));
-}
-
-#[test]
-fn test_infer_call_expression_type_deduplicates_substituted_union_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function maybeString<T>(value: T): T | string {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let maybe_string_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "maybeString")
-        .expect("maybeString binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, maybe_string_ty),
-        Vec::from([InferredTypeData::String]),
-    );
-
-    assert!(is_inferred_string(&db, call_ty));
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_deduplicates_substituted_union_return_type",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_call_expression_type_substitutes_callback_generic_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function run<T>(cb: () => T): T {
-                return cb();
-            }
-
-            export function readString(): string {
-                return "value";
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let run_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "run")
-        .expect("run binding type must be inferred");
-    let read_string_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "readString")
-        .expect("readString binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, run_ty),
-        Vec::from([inferred.resolve_type(&db, read_string_ty)]),
-    );
-
-    assert!(is_inferred_string(&db, call_ty));
-}
-
-#[test]
-fn test_infer_call_expression_type_substitutes_nested_generic_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function wrap<T>(value: T): Promise<T> {
-                return Promise.resolve(value);
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let wrap_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "wrap")
-        .expect("wrap binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, wrap_ty),
-        Vec::from([InferredTypeData::Number]),
-    );
-    let InferredTypeData::InstanceOf(instance) = call_ty else {
-        panic!("wrap must return a Promise instance, got {call_ty:?}");
-    };
-
-    assert!(call_ty.is_promise_instance(&db));
-    assert!(
-        instance
-            .type_parameters(&db)
-            .iter()
-            .any(|ty| is_inferred_number(&db, *ty))
-    );
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_substitutes_nested_generic_return_type",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_call_expression_type_substitutes_generic_inside_promise_union_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function boxed<T>(value: T): Promise<T | number> {
-                return Promise.resolve(value as T | number);
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let boxed_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "boxed")
-        .expect("boxed binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, boxed_ty),
-        Vec::from([InferredTypeData::String]),
-    );
-
-    assert!(is_inferred_promise_with_type_parameter(
-        &db,
-        call_ty,
-        |ty| { contains_inferred_string(&db, ty) && contains_inferred_number(&db, ty) }
-    ));
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_substitutes_generic_inside_promise_union_return_type",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_call_expression_type_substitutes_multiple_generics_inside_union_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function pair<T, U>(left: T, right: U): Promise<T | U> {
-                return Promise.resolve(left as T | U);
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let pair_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "pair")
-        .expect("pair binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, pair_ty),
-        Vec::from([InferredTypeData::String, InferredTypeData::Number]),
-    );
-
-    assert!(is_inferred_promise_with_type_parameter(
-        &db,
-        call_ty,
-        |ty| { contains_inferred_string(&db, ty) && contains_inferred_number(&db, ty) }
-    ));
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_substitutes_multiple_generics_inside_union_return_type",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_call_expression_type_substitutes_generic_inside_union_with_promise_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function maybePromise<T>(value: T): T | Promise<T> {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let maybe_promise_ty =
-        inferred_binding_ty_by_name(&db, index_module, &inferred, "maybePromise")
-            .expect("maybePromise binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, maybe_promise_ty),
-        Vec::from([InferredTypeData::String]),
-    );
-    let InferredTypeData::Union(union) = call_ty else {
-        panic!("maybePromise must return a union, got {call_ty:?}");
-    };
-
-    assert!(
-        union
-            .types(&db)
-            .iter()
-            .any(|ty| contains_inferred_string(&db, *ty))
-    );
-    assert!(union.types(&db).iter().any(|ty| {
-        is_inferred_promise_with_type_parameter(&db, *ty, |ty| contains_inferred_string(&db, ty))
-    }));
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_substitutes_generic_inside_union_with_promise_return_type",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_call_expression_type_substitutes_generic_inside_intersection_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            type WithName = {
-                name: string;
-            };
-
-            export function withName<T>(value: T): T & WithName {
-                return value as T & WithName;
-            }
-
-            export function readValue(): { value: number } {
-                return { value: 1 };
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let with_name_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "withName")
-        .expect("withName binding type must be inferred");
-    let read_value_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readValue")
-            .expect("readValue return type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, with_name_ty),
-        Vec::from([read_value_ty]),
-    );
-    let InferredTypeData::Object(object) = call_ty else {
-        panic!("withName call must return a normalized object, got {call_ty:?}");
-    };
-    assert_eq!(object.members(&db).len(), 2);
-
-    let name_ty = inferred
-        .find_member_type(&db, call_ty, "name")
-        .expect("normalized call return must expose WithName.name");
-    assert!(is_inferred_string(&db, name_ty));
-
-    let value_ty = inferred
-        .find_member_type(&db, call_ty, "value")
-        .expect("normalized call return must expose readValue.value");
-    assert!(is_inferred_number(&db, value_ty));
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_substitutes_generic_inside_intersection_return_type",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_call_expression_type_substitutes_generic_from_callback_promise_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function unwrap<T>(cb: () => Promise<T>): T {
-                return undefined as T;
-            }
-
-            export function readNumber(): Promise<number> {
-                return Promise.resolve(1);
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let unwrap_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "unwrap")
-        .expect("unwrap binding type must be inferred");
-    let read_number_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "readNumber")
-        .expect("readNumber binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, unwrap_ty),
-        Vec::from([inferred.resolve_type(&db, read_number_ty)]),
-    );
-
-    assert!(is_inferred_number(&db, call_ty));
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_substitutes_generic_from_callback_promise_return_type",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_call_expression_type_resolves_union_function_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function readString(): string {
-                return "value";
-            }
-
-            export function readNumber(): number {
-                return 1;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let read_string_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "readString")
-        .expect("readString binding type must be inferred");
-    let read_number_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "readNumber")
-        .expect("readNumber binding type must be inferred");
-    let callee_ty = InferredTypeData::Union(InferredUnion::new(
-        &db,
-        Vec::from([
-            inferred.resolve_type(&db, read_string_ty),
-            inferred.resolve_type(&db, read_number_ty),
-        ])
-        .into_boxed_slice(),
-    ));
-    let call_ty = infer_call_expression_type(&db, index_module, callee_ty, Vec::new());
-    let InferredTypeData::Union(union) = call_ty else {
-        panic!("union function call must return a union, got {call_ty:?} from {callee_ty:?}");
-    };
-
-    assert!(
-        union
-            .types(&db)
-            .iter()
-            .any(|ty| is_inferred_string(&db, *ty))
-    );
-    assert!(
-        union
-            .types(&db)
-            .iter()
-            .any(|ty| is_inferred_number(&db, *ty))
-    );
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_resolves_union_function_return_type",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_module_types_normalizes_union_variants_on_build() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function readValue(value: string | (number | string)): string | (number | string) {
-                return value;
-            }
-
-            export function readBoolean(value: true | false): true | false {
-                return value;
-            }
-
-            export function readStringLiteral(value: string | "literal"): string | "literal" {
-                return value;
-            }
-
-            export function readNumberLiteral(value: number | 1): number | 1 {
-                return value;
-            }
-
-            export function readBigIntLiteral(value: bigint | 1n): bigint | 1n {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let value_ty = inferred_function_return_ty_by_name(&db, index_module, &inferred, "readValue")
-        .expect("readValue return type must be inferred");
-    let InferredTypeData::Union(union) = value_ty else {
-        panic!("readValue must return a union, got {value_ty:?}");
-    };
-    assert_eq!(union.types(&db).len(), 2);
-    assert!(
-        union
-            .types(&db)
-            .iter()
-            .all(|ty| !matches!(ty, InferredTypeData::Union(_)))
-    );
-    assert!(
-        union
-            .types(&db)
-            .iter()
-            .any(|ty| is_inferred_string(&db, *ty))
-    );
-    assert!(
-        union
-            .types(&db)
-            .iter()
-            .any(|ty| is_inferred_number(&db, *ty))
-    );
-
-    let boolean_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readBoolean")
-            .expect("readBoolean return type must be inferred");
-    assert_eq!(boolean_ty, InferredTypeData::Boolean);
-
-    let string_literal_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readStringLiteral")
-            .expect("readStringLiteral return type must be inferred");
-    assert_eq!(string_literal_ty, InferredTypeData::String);
-
-    let number_literal_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readNumberLiteral")
-            .expect("readNumberLiteral return type must be inferred");
-    assert_eq!(number_literal_ty, InferredTypeData::Number);
-
-    let bigint_literal_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readBigIntLiteral")
-            .expect("readBigIntLiteral return type must be inferred");
-    assert_eq!(bigint_literal_ty, InferredTypeData::BigInt);
-
-    assert_inferred_type_snapshot(
-        "test_infer_module_types_normalizes_union_variants_on_build",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_module_types_normalizes_intersection_variants_on_build() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            type WithName = {
-                name: string;
-            };
-
-            type WithValue = {
-                value: number;
-            };
-
-            export function readCombined(
-                value: WithName & (WithValue & WithName),
-            ): WithName & (WithValue & WithName) {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let combined_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readCombined")
-            .expect("readCombined return type must be inferred");
-    let InferredTypeData::Intersection(intersection) = combined_ty else {
-        panic!("readCombined must return an intersection, got {combined_ty:?}");
-    };
-    assert_eq!(intersection.types(&db).len(), 2);
-    assert!(
-        intersection
-            .types(&db)
-            .iter()
-            .all(|ty| !matches!(ty, InferredTypeData::Intersection(_)))
-    );
-
-    let name_ty = inferred
-        .find_member_type(&db, combined_ty, "name")
-        .expect("normalized intersection must expose WithName.name");
-    assert!(is_inferred_string(&db, name_ty));
-
-    let value_ty = inferred
-        .find_member_type(&db, combined_ty, "value")
-        .expect("normalized intersection must expose WithValue.value");
-    assert!(is_inferred_number(&db, value_ty));
-
-    let normalized_ty = normalize_type(&db, index_module, combined_ty);
-    let InferredTypeData::Object(normalized_object) = normalized_ty else {
-        panic!("normalized Local intersection must become an object, got {normalized_ty:?}");
-    };
-    assert_eq!(normalized_object.members(&db).len(), 2);
-
-    let normalized_name_ty = inferred
-        .find_member_type(&db, normalized_ty, "name")
-        .expect("normalized object must expose WithName.name");
-    assert!(is_inferred_string(&db, normalized_name_ty));
-
-    let normalized_value_ty = inferred
-        .find_member_type(&db, normalized_ty, "value")
-        .expect("normalized object must expose WithValue.value");
-    assert!(is_inferred_number(&db, normalized_value_ty));
-
-    assert_inferred_type_snapshot(
-        "test_infer_module_types_normalizes_intersection_variants_on_build",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_module_types_normalizes_primitive_intersections_on_build() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function readString(value: string & string): string & string {
-                return value;
-            }
-
-            export function readNever(value: string & number): string & number {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let string_ty = inferred_function_return_ty_by_name(&db, index_module, &inferred, "readString")
-        .expect("readString return type must be inferred");
-    assert_eq!(string_ty, InferredTypeData::String);
-
-    let never_ty = inferred_function_return_ty_by_name(&db, index_module, &inferred, "readNever")
-        .expect("readNever return type must be inferred");
-    assert_eq!(never_ty, InferredTypeData::NeverKeyword);
-
-    assert_inferred_type_snapshot(
-        "test_infer_module_types_normalizes_primitive_intersections_on_build",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_module_types_merges_inline_object_intersections_on_build() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function readCombined(
-                value: { name: string } & { value: number },
-            ): { name: string } & { value: number } {
-                return value;
-            }
-
-            export function readValue(
-                value: { value: string } & { value: number },
-            ): { value: string } & { value: number } {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let combined_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readCombined")
-            .expect("readCombined return type must be inferred");
-    let InferredTypeData::Object(combined_object) = combined_ty else {
-        panic!("readCombined must return a merged object, got {combined_ty:?}");
-    };
-    assert_eq!(combined_object.members(&db).len(), 2);
-
-    let name_ty = inferred
-        .find_member_type(&db, combined_ty, "name")
-        .expect("merged object must expose name");
-    assert!(is_inferred_string(&db, name_ty));
-
-    let value_ty = inferred
-        .find_member_type(&db, combined_ty, "value")
-        .expect("merged object must expose value");
-    assert!(is_inferred_number(&db, value_ty));
-
-    let duplicate_value_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readValue")
-            .expect("readValue return type must be inferred");
-    let duplicate_value_member_ty = inferred
-        .find_member_type(&db, duplicate_value_ty, "value")
-        .expect("merged duplicate member must expose value");
-    assert!(contains_inferred_string(&db, duplicate_value_member_ty));
-    assert!(contains_inferred_number(&db, duplicate_value_member_ty));
-
-    assert_inferred_type_snapshot(
-        "test_infer_module_types_merges_inline_object_intersections_on_build",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_module_types_merges_function_intersections_on_build() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function readFunction(
-                value: (() => string) & (() => number),
-            ): (() => string) & (() => number) {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let function_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readFunction")
-            .expect("readFunction return type must be inferred");
-    let InferredTypeData::Function(function) = function_ty else {
-        panic!("readFunction must return a merged function, got {function_ty:?}");
-    };
-    let InferredReturnType::Type(return_ty) = function.return_type(&db) else {
-        panic!("merged function return type must be a type");
-    };
-
-    assert!(contains_inferred_string(&db, *return_ty));
-    assert!(contains_inferred_number(&db, *return_ty));
-
-    assert_inferred_type_snapshot(
-        "test_infer_module_types_merges_function_intersections_on_build",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_module_types_merges_mixed_intersections_on_build() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function readCallableObject(
-                value: (() => string) & { value: number },
-            ): (() => string) & { value: number } {
-                return value;
-            }
-
-            export function readPrimitive(
-                value: string & { value: number },
-            ): string & { value: number } {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let callable_object_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readCallableObject")
-            .expect("readCallableObject return type must be inferred");
-    let InferredTypeData::Object(callable_object) = callable_object_ty else {
-        panic!("readCallableObject must return a merged object, got {callable_object_ty:?}");
-    };
-    assert_eq!(callable_object.members(&db).len(), 1);
-    let value_ty = inferred
-        .find_member_type(&db, callable_object_ty, "value")
-        .expect("merged callable object must expose value");
-    assert!(is_inferred_number(&db, value_ty));
-
-    let primitive_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readPrimitive")
-            .expect("readPrimitive return type must be inferred");
-    assert_eq!(primitive_ty, InferredTypeData::String);
-
-    assert_inferred_type_snapshot(
-        "test_infer_module_types_merges_mixed_intersections_on_build",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_module_types_merges_class_instance_intersections_on_build() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function readPromiseObject(
-                value: Promise<string> & { value: number },
-            ): Promise<string> & { value: number } {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let promise_object_ty =
-        inferred_function_return_ty_by_name(&db, index_module, &inferred, "readPromiseObject")
-            .expect("readPromiseObject return type must be inferred");
-    let InferredTypeData::InstanceOf(instance) = promise_object_ty else {
-        panic!("readPromiseObject must return a merged class instance, got {promise_object_ty:?}");
-    };
-    assert!(matches!(instance.ty(&db), InferredTypeData::Class(_)));
-
-    let value_ty = inferred
-        .find_member_type(&db, promise_object_ty, "value")
-        .expect("merged class instance must expose value");
-    assert!(is_inferred_number(&db, value_ty));
-
-    assert_inferred_type_snapshot(
-        "test_infer_module_types_merges_class_instance_intersections_on_build",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_module_types_resolves_imported_local_handle_members() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/types.ts".into(),
-        r#"
-            export class Foo {
-                name: string;
-
-                static create(): Foo {
-                    return new Foo();
-                }
-            }
-        "#,
-    );
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            import { Foo } from "./types.ts";
-
-            export const value: Foo = Foo.create();
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/types.ts", "/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let foo_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "Foo")
-        .expect("Foo import type must be inferred");
-
-    let create_ty = inferred
-        .find_member_type(&db, foo_ty, "create")
-        .expect("Foo.create must be found through the imported local handle");
-    let InferredTypeData::Function(create_function) = create_ty else {
-        panic!("Foo.create must be a function");
-    };
-    let InferredReturnType::Type(return_ty) = create_function.return_type(&db) else {
-        panic!("Foo.create return type must be a type");
-    };
-
-    let name_ty = inferred
-        .find_member_type(&db, *return_ty, "name")
-        .expect("Foo.create().name must be found through the imported local handle");
-    assert!(is_inferred_string(&db, name_ty));
-}
-
-#[test]
-fn test_infer_module_types_resolves_namespace_import_members() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/source.ts".into(),
-        r#"
-            export function alpha(): number {
-                return 1;
-            }
-        "#,
-    );
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            import * as source from "./source.ts";
-
-            export { source };
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/source.ts", "/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let source_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "source")
-        .expect("source import type must be inferred");
-    let InferredTypeData::Namespace(_) = source_ty else {
-        panic!("namespace import must infer a namespace, got {source_ty:?}");
-    };
-
-    let alpha_ty = inferred
-        .find_member_type(&db, source_ty, "alpha")
-        .expect("namespace import must expose source.alpha");
-    assert_inferred_function_returns_number(&db, alpha_ty);
-
-    assert_inferred_type_snapshot(
-        "test_infer_module_types_resolves_namespace_import_members",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_module_types_resolves_namespace_reexport_members() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/source.ts".into(),
-        r#"
-            export function alpha(): number {
-                return 1;
-            }
-        "#,
-    );
-    fs.insert(
-        "/src/barrel.ts".into(),
-        r#"export * as MyNs from "./source.ts";"#,
-    );
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            import { MyNs } from "./barrel.ts";
-
-            export { MyNs };
-        "#,
-    );
-
-    let db = build_js_test_module_db(
-        &fs,
-        &["/src/source.ts", "/src/barrel.ts", "/src/index.ts"],
-        true,
-    );
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let namespace_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "MyNs")
-        .expect("MyNs import type must be inferred");
-    let InferredTypeData::Namespace(_) = namespace_ty else {
-        panic!("namespace reexport import must infer a namespace, got {namespace_ty:?}");
-    };
-
-    let alpha_ty = inferred
-        .find_member_type(&db, namespace_ty, "alpha")
-        .expect("namespace reexport must expose source.alpha");
-    assert_inferred_function_returns_number(&db, alpha_ty);
-
-    assert_inferred_type_snapshot(
-        "test_infer_module_types_resolves_namespace_reexport_members",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_module_types_resolves_imported_inherited_members() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/base.ts".into(),
-        r#"
-            export class Base {
-                name: string;
-
-                static label(): string {
-                    return "base";
-                }
-            }
-        "#,
-    );
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            import { Base } from "./base.ts";
-
-            export class Derived extends Base {
-                value: number;
-            }
-
-            export const derived: Derived = new Derived();
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/base.ts", "/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let derived_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "derived")
-        .expect("derived binding type must be inferred");
-    let name_ty = inferred
-        .find_member_type(&db, derived_ty, "name")
-        .expect("Derived instance must inherit imported Base.name");
-    assert!(is_inferred_string(&db, name_ty));
-
-    let derived_class_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "Derived")
-        .expect("Derived class type must be inferred");
-    let label_ty = inferred
-        .find_member_type(&db, derived_class_ty, "label")
-        .expect("Derived class must inherit imported Base.label");
-    assert!(matches!(label_ty, InferredTypeData::Function(_)));
-}
-
-#[test]
-fn test_infer_module_types_resolves_imported_interface_extends_members() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/base.ts".into(),
-        r#"
-            export interface Base {
-                name: string;
-            }
-        "#,
-    );
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            import type { Base } from "./base.ts";
-
-            export interface Derived extends Base {
-                value: number;
-            }
-
-            export const derived: Derived = {
-                name: "derived",
-                value: 1,
-            };
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/base.ts", "/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let derived_ty = inferred_binding_ty_by_name(&db, index_module, &inferred, "derived")
-        .expect("derived binding type must be inferred");
-
-    let name_ty = inferred
-        .find_member_type(&db, derived_ty, "name")
-        .expect("Derived interface must inherit imported Base.name");
-    assert!(is_inferred_string(&db, name_ty));
-}
-
-#[test]
-fn test_infer_module_types_is_memoized() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export const value: string = "value";
-        "#,
-    );
-
-    let mut db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-
-    let _ = infer_module_types(&db, index_module);
-    db.clear_salsa_events();
-    let _ = infer_module_types(&db, index_module);
-    let events = db.take_salsa_events();
-
-    assert_function_query_was_not_run(&db, infer_module_types, index_module, &events);
-}
-
-#[test]
-fn test_infer_module_types_backdates_equal_output() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export const value: string = "value";
-        "#,
-    );
-
-    let mut db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let expression_count = inferred_expression_count(&db, index_module);
-    assert!(expression_count > 0);
-
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export const value: string = "value";
-            // This changes the module input, but not the inferred types.
-        "#,
-    );
-    let module_kind = resolve_js_module_kind_for_test(&fs, "/src/index.ts", true);
-    salsa::Setter::to(index_module.set_kind(&mut db), module_kind);
-
-    db.clear_salsa_events();
-    assert_eq!(
-        inferred_expression_count(&db, index_module),
-        expression_count
-    );
-    let events = db.take_salsa_events();
-
-    assert_function_query_was_run(&db, infer_module_types, index_module, &events);
-    assert_function_query_was_not_run(&db, inferred_expression_count, index_module, &events);
-}
-
-#[test]
-fn test_infer_module_types_bottom_up_warms_blanket_reexports() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/leaf.ts".into(),
-        r#"
-            export type Source = {
-                name: string;
-            };
-        "#,
-    );
-    fs.insert(
-        "/src/mid.ts".into(),
-        r#"
-            export * from "./leaf.ts";
-        "#,
-    );
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            import type { Source } from "./mid.ts";
-
-            export function read(value: Source): Source {
-                return value;
-            }
-        "#,
-    );
-
-    let mut db =
-        build_js_test_module_db(&fs, &["/src/leaf.ts", "/src/mid.ts", "/src/index.ts"], true);
-    let leaf_module = db
-        .module_for_path(Utf8Path::new("/src/leaf.ts"))
-        .expect("leaf module must exist");
-    let mid_module = db
-        .module_for_path(Utf8Path::new("/src/mid.ts"))
-        .expect("mid module must exist");
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("index module must exist");
-
-    db.clear_salsa_events();
-    let inferred = infer_module_types_bottom_up(&db, index_module).expect("types must be inferred");
-    let return_ty = inferred_function_return_ty_by_name(&db, index_module, &inferred, "read")
-        .expect("read return type must be inferred");
-    let name_ty = inferred
-        .find_member_type(&db, return_ty, "name")
-        .expect("re-exported Source must expose name");
-    assert!(is_inferred_string(&db, name_ty));
-
-    let events = db.take_salsa_events();
-    let leaf_position =
-        function_query_will_execute_position(&db, infer_module_types, leaf_module, &events)
-            .expect("leaf inference must run");
-    let mid_position =
-        function_query_will_execute_position(&db, infer_module_types, mid_module, &events)
-            .expect("mid inference must run");
-    let index_position =
-        function_query_will_execute_position(&db, infer_module_types, index_module, &events)
-            .expect("index inference must run");
-    assert!(
-        leaf_position < mid_position && mid_position < index_position,
-        "bottom-up inference must warm blanket re-export dependencies before their importers"
-    );
 }
 
 #[test]
